streaming/feature_store/requirements.txt
# Feature Store Dependencies
redis==5.0.1
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.24.3
pyyaml==6.0.1
scipy==1.11.4
# Для асинхронного сервинга
fastapi==0.104.1
uvicorn[standard]==0.24.0
# Для мониторинга
prometheus-client==0.19.0